% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdf.R
\name{as_ctdf_track}
\alias{as_ctdf_track}
\title{Convert a \code{ctdf} track to movement step segments as LINESTRINGs}
\usage{
as_ctdf_track(ctdf)
}
\arguments{
\item{ctdf}{A \code{ctdf} object (with ordered rows and a \code{"location"} geometry column).}
}
\value{
An \code{sf} object with LINESTRING geometry for each step.
}
\description{
Takes a \code{ctdf} object and returns an \code{sf} object with LINESTRING geometries representing
the movement steps between consecutive locations. Each segment connects two points,
starting at the previous location and ending at the current one - i.e., each segment
ends at the position of the current row.
}
\details{
The number of rows is nrow(ctdf) - i, where i = 1 and corresponds to the starting index in ctdf.
}
\examples{
data(mini_ruff)
ctdf = as_ctdf(mini_ruff)
s = as_ctdf_track(ctdf)
plot(s['.id'])

}
