% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_clean.R
\name{clean_ctdf}
\alias{clean_ctdf}
\title{Clean putative clusters after slicing}
\usage{
clean_ctdf(ctdf, Q = 0.95)
}
\arguments{
\item{ctdf}{A cluster-track data.table after slicing}
}
\value{
A cleaned cluster-track data.table containing only refined
  putative cluster points
}
\description{
Slicing removes selected movement segments from a track and identifies
putative cluster regions. At this stage, putative clusters are not yet
valid clusters because they may still contain track fragments and outliers.
This function removes outliers and residual movement segments from
putative clusters.
}
