% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTrack-package.R
\docType{package}
\name{clusterTrack-package}
\alias{clusterTrack}
\alias{clusterTrack-package}
\title{clusterTrack: Spatiotemporal clustering of animal telemetry tracks}
\description{
\code{clusterTrack} identifies spatiotemporally distinct use sites from animal
telemetry tracks.
}
\section{Package options}{

\code{clusterTrack} uses the following options:

\code{clusterTrack.verbose}: Logical. Default \code{TRUE}. Set to \code{FALSE} to
suppress \code{cli} alerts emitted during package workflows.

\code{clusterTrack.max_gap}: Numeric, in hours. Default \code{24}. A warning reports
adjacent timestamps separated by more than this threshold.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(
  clusterTrack.verbose = FALSE,
  clusterTrack.max_gap = 24
)
}\if{html}{\out{</div>}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://xarvian.github.io/clusterTrack/}
  \item Report bugs at \url{https://github.com/xarvian/clusterTrack/issues}
}

}
\author{
\strong{Maintainer}: Anonymous Author \email{293027538+xarvian@users.noreply.github.com}

Authors:
\itemize{
  \item Anonymous Author \email{293027538+xarvian@users.noreply.github.com}
}

}
\keyword{internal}
