% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_outliers.R
\name{ctdf_lof}
\alias{ctdf_lof}
\title{Compute cluster-wise local outlier factor scores with \code{dbscan::lof()}}
\usage{
ctdf_lof(ctdf, minPts = NULL)
}
\arguments{
\item{ctdf}{A \code{ctdf} object .}

\item{minPts}{Optional integer passed to \code{\link[dbscan:lof]{dbscan::lof()}}. If \code{NULL},
\code{\link[dbscan:lof]{dbscan::lof()}} is called with its defaults.

The function updates \code{ctdf} by reference by updating the \code{lof} column.}
}
\description{
Uses \code{\link[dbscan:lof]{dbscan::lof()}} to compute Local Outlier Factor (LOF) scores
separately within each \code{cluster} of a \code{ctdf}, and writes the
result into the \code{lof} column.
}
\examples{
data(mini_ruff)
x = as_ctdf(mini_ruff)
cluster_track(x)
x = ctdf_lof(x)
head(x[, .(.id, cluster, lof)])
}
\seealso{
\code{\link[dbscan:lof]{dbscan::lof()}}, \code{\link[=cluster_track]{cluster_track()}}.
}
