% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{.cutoff_by_mixfit}
\alias{.cutoff_by_mixfit}
\title{Find the cutoff that maximizes expected classification into component 1}
\usage{
.cutoff_by_mixfit(v)
}
\description{
Find the cutoff that maximizes expected classification into component 1
}
