% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_repair.R
\name{.is_intersection}
\alias{.is_intersection}
\title{Repair temporally adjacent clusters}
\usage{
.is_intersection(ctdf, pc, next_pc)
}
\arguments{
\item{ctdf}{A CTDF object. Must contain an updated `cluster` column.}
}
\value{
The input CTDF, with an updated (in-place) cluster` column.
}
\description{
Merge spatial clusters.
}
\examples{
data(mini_ruff)
ctdf = as_ctdf(mini_ruff)
slice_ctdf(ctdf)
spatial_repair(ctdf)

}
