% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_ctdf}
\alias{filter_ctdf}
\title{Top-Down Noise Pruning for Tracks}
\usage{
filter_ctdf(ctdf, method = c("intersection", "density_gap"), ...)
}
\arguments{
\item{method}{Character; which pruning strategy to apply. One of:
- `"intersection"` : remove non-intersecting segments (and optionally their neighbors)  
- `"speed"`        : a simple speed filter.}

\item{...}{Additional arguments passed to the chosen methods:}

\item{x}{An object of class **ctdf** (from [as_ctdf()])}
}
\value{
A  logical vector the same length as x.
}
\description{
A generic wrapper which flags segments by cutting out
}
\details{
This function implements a **top-down** noise-pruning step, splitting
away low-density or isolated portions of your track graph before any
downstream clustering.
}
\examples{
library(clusterTrack)
data(pesa56511)
ctdf = as_ctdf(pesa56511, time="locationDate")


}
\seealso{
- [as_ctdf()]       for converting `data.table` → `ctdf`  
- [track_segments()] for building an `sf` of segments from a `ctdf`  
- `fsr_intersection()`,
}
