% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_intersection}
\alias{filter_intersection}
\title{Filter by Segment Intersection}
\usage{
filter_intersection(ctdf, overwrite = FALSE, strict = TRUE)
}
\arguments{
\item{ctdf}{An sf object whose geometries are line‐segments in sequence.}

\item{strict}{If TRUE (the default), only segments with both neighbors also isolated are kept.
               If FALSE, any isolated segment is kept.
#' @param overwrite Logical scalar, default FALSE. If FALSE, the existing filter column is updated by marking any new points to filter; if TRUE, the filter column is replaced entirely.}
}
\value{
A  logical vector the same length as ctdf.
}
\description{
Flags segments that are isolated (or isolated within their immediate neighbors).
}
\examples{
data(zbird)
ctdf = as_ctdf(zbird)
filter_intersection(ctdf)

}
