% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{fsr_intersection}
\alias{fsr_intersection}
\title{Top-Down Noise Pruning by Segment Intersection}
\usage{
fsr_intersection(ctdf, strict = TRUE)
}
\arguments{
\item{ctdf}{An sf object whose geometries are line‐segments in sequence.}

\item{strict}{If TRUE, only segments with both neighbors also isolated are kept.
If FALSE, any isolated segment is kept.}
}
\value{
A  logical vector the same length as ctdf.
}
\description{
Marks segments that are isolated (or isolated within their immediate neighbors).
}
\examples{
library(clusterTrack)
data(pesa56511)
ctdf = as_ctdf(pesa56511, time="locationDate")

}
