% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbdo66867}
\alias{lbdo66867}
\title{ARGOS satellite tracking data for an individual Long-billed dowitcher}
\format{
A data.table with 633 rows and 4 columns:
\describe{
  \item{latitude}{Numeric. Latitude in decimal degrees (WGS84).}
  \item{longitude}{Numeric. Longitude in decimal degrees (WGS84).}
  \item{locationDate}{POSIXct. Timestamp of location fix (UTC).}
  \item{locationClass}{Character. ARGOS location quality class (e.g., B, 0, A).}
}
}
\source{
Department of Ornithology, Max Planck Institute for Biological Intelligence.
}
\usage{
lbdo66867
}
\description{
A dataset containing ARGOS locations for an individual male Long-billed dowitcher.
}
\examples{
data(lbdo66867)
head(lbdo66867)
}
\keyword{datasets}
