% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTrack.Vis.R
\name{map}
\alias{map}
\title{Map a ctdf}
\usage{
map(ctdf)
}
\description{
Map a ctdf
}
\examples{
data(pesa56511)
ctdf  = as_ctdf(pesa56511, time = "locationDate", crs = 4326, project_to = "+proj=eqearth") |>cluster_track()
map(ctdf)
}
