% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nola125a}
\alias{nola125a}
\title{GNSS tracking data for an individual northern lapwing.}
\format{
A data.table with 2484 rows and 3 columns:
\describe{
\item{timestamp}{POSIXct. Timestamp of location fix (UTC).}
\item{latitude}{Numeric. Latitude in decimal degrees (WGS84).}
\item{longitude}{Numeric. Longitude in decimal degrees (WGS84).}
}
}
\source{
Unpublished data.
}
\usage{
nola125a
}
\description{
A dataset containing GNSS locations for an individual female northern lapwing.
}
\examples{
data(nola125a)
head(nola125a)
}
\keyword{datasets}
