% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_outliers.R
\name{outliers}
\alias{outliers}
\title{Score outliers in a ctdf using dbscan methods}
\usage{
outliers(
  ctdf,
  method = c("lof", "glosh"),
  type = c("within-cluster", "between-clusters"),
  minPts,
  k
)
}
\arguments{
\item{ctdf}{A \code{ctdf} object.}

\item{method}{Character scalar. One of \code{"lof"} or \code{"glosh"}.}

\item{type}{Character scalar. Currently only \code{"within-cluster"} is supported.}

\item{minPts}{Optional integer. Passed to \code{\link[dbscan:lof]{dbscan::lof()}}.}

\item{k}{Optional integer. Passed to \code{\link[dbscan:glosh]{dbscan::glosh()}}.}
}
\value{
A copy of \code{ctdf} with one appended outlier score column.
}
\description{
Computes outlier scores for each location in a \code{ctdf} and returns a copy
with one appended score column.
}
\details{
For now, only within-cluster scoring is supported. Existing \code{cluster}
labels are used only to define subsets. Outlier scores are then computed
independently within each subset.
}
