% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdf.R
\docType{data}
\name{reserved_ctdf_nams}
\alias{reserved_ctdf_nams}
\title{Reserved ctdf column names}
\format{
An object of class \code{character} of length 6.
}
\usage{
reserved_ctdf_nams
}
\description{
Reserved ctdf column names
}
\keyword{internal}
