% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ruff07b5}
\alias{ruff07b5}
\title{GNSS tracking data for an individual ruff.}
\format{
A data.table with 2772 rows and 3 columns:
\describe{
\item{timestamp}{POSIXct. Timestamp of location fix (UTC).}
\item{latitude}{Numeric. Latitude in decimal degrees (WGS84).}
\item{longitude}{Numeric. Longitude in decimal degrees (WGS84).}
}
}
\source{
Unpublished data.
}
\usage{
ruff07b5
}
\description{
A dataset containing GNSS locations for an individual male ruff.
}
\examples{
data(ruff07b5)
head(ruff07b5)
}
\keyword{datasets}
