% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment_ctdf}
\alias{segment_ctdf}
\title{segment ctdf
segment_ctdf}
\usage{
segment_ctdf(ctdf, deltaT = 12, min_n_segments = 5, bout_length)
}
\description{
segment ctdf
segment_ctdf
}
\examples{
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2 )
se = segment_ctdf(ctdf, deltaT = 12) 
mapview(se, zcol = "bout_id")

}
