% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_slice.R
\name{slice_ctdf}
\alias{slice_ctdf}
\title{Slice a CTDF into putative clusters using temporal continuity and spatial clustering}
\usage{
slice_ctdf(ctdf, nmin = 5, deltaT)
}
\arguments{
\item{ctdf}{A \emph{CTDF} object.}

\item{nmin}{Integer;  smallest size of a putative cluster.}

\item{deltaT}{Numeric; maximum allowable time gap (in days) between segment endpoints for intersections
to  consider them continuous.}
}
\value{
Invisibly returns \code{ctdf}, with \code{.putative_cluster} updated in-place.
}
\description{
Identifies spatially heterogeneous regions (via HDBSCAN on point coordinates) and, for those regions,
recursively subdivides the track into temporally continuous movement segments. Subdivision continues
until a region is spatially homogeneous (no evidence for multiple clusters) .
}
\details{
This function updates a \code{ctdf} in-place.

Internally, candidate regions are queued. Regions that show evidence for multiple clusters are split by movement segmentation; otherwise they are retained as a single putative cluster.
}
\examples{
data(mini_ruff)
ctdf = as_ctdf(mini_ruff, s_srs = 4326, t_srs = "+proj=eqearth")
ctdf = slice_ctdf(ctdf)
}
\seealso{
\code{\link[dbscan]{hdbscan}}
}
