% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch_cluster}
\alias{stitch_cluster}
\title{stitch_cluster
stitch a cluster}
\usage{
stitch_cluster(clust, dT = 4, overlap_threshold = 0.5)
}
\description{
stitch_cluster
stitch a cluster
}
\examples{
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2, crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf)
clust = cluster_segments(ctdf )
data(pesa56511)
ctdf  = as_ctdf(pesa56511, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf) 
clust = cluster_segments(ctdf) |> stitch_cluster(dT = 4000 )
map(clust)
}
