% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{tdbscan}
\alias{tdbscan}
\title{tdbscan}
\usage{
tdbscan(track, eps, minPts = 5, borderPoints = FALSE, maxLag = 6, minTenure)
}
\arguments{
\item{track}{A sf object (for now).}

\item{eps}{size of the epsilon neighborhood (see \code{\link[dbscan:dbscan]{dbscan::dbscan()}}  ).}

\item{minPts}{number of minimum points in the eps region (for core points).
Default is 5 points (see \code{\link[dbscan:dbscan]{dbscan::dbscan()}}  ).}

\item{borderPoints}{Logical; default to FALSE (see \code{\link[dbscan:dbscan]{dbscan::dbscan()}}  ).}

\item{maxLag}{maximum relative temporal lag (see notes). Default to 6.}

\item{minTenure}{minimum time difference, in hours, between the last and the first entry a cluster.
Clusters with  values smaller than minTenure are discarded.}
}
\description{
tdbscan
}
\note{
When maxLag is set to \code{maxLag>N} output is the same as for \link[dbscan:dbscan]{dbscan}.
}
\examples{
data(pesa56511)
x = as_tdbscan(pesa56511, time = "locationDate", s_srs = 4326)
x = st_transform(x, '+proj=eqearth')
z = tdbscan(track=x, eps =6600 , minPts   = 8, maxLag = 6, borderPoints = TRUE )

# Set minTenure
z = tdbscan(x, eps =6600, minPts   = 8, maxLag = 6, borderPoints = TRUE, minTenure= 24 )

}
