% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_ctdf}
\alias{aggregate_ctdf}
\title{Aggregate ctdf data}
\usage{
aggregate_ctdf(ctdf, timescale = "1 day")
}
\arguments{
\item{ctdf}{A `ctdf` object.}

\item{timescale}{Time aggregation scale, default is "1 day".}
}
\description{
Aggregate ctdf data
}
