% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_ctdf}
\alias{aggregate_ctdf}
\title{Aggregate ctdf data}
\usage{
aggregate_ctdf(ctdf, dist)
}
\arguments{
\item{ctdf}{A `ctdf` object.}

\item{dist}{aggregation scale}
}
\description{
Aggregate ctdf data
}
