% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdf.R
\name{as_ctdf}
\alias{as_ctdf}
\title{Coerce a data.table to clusterTrack data format}
\usage{
as_ctdf(x, coords = c("longitude", "latitude"), time = "time", crs = 4326, ...)
}
\arguments{
\item{x}{A `data.table` object.}

\item{coords}{Character vector of length 2 specifying the coordinate column names.
Defaults to `c("longitude", "latitude")`.}

\item{time}{Name of the time column. Will be renamed to `"timestamp"` internally.}

\item{crs}{Coordinate reference system. Default is EPSG 4326.}

\item{...}{Currently unused}
}
\value{
An object of class `ctdf`, which inherits from `sf`.
}
\description{
Converts a `data.table` with spatial coordinates and a timestamp column
to a standardized `sf`-based format used internally by the clusterTrack package.
}
\note{
This is currently a thin wrapper around `st_as_sf()`, but standardizes timestamp naming, ordering,
and geometry column name (`"location"`).
}
\examples{
data(zbird)
x <- as_ctdf(zbird)

}
