% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdf.R
\name{as_ctdf}
\alias{as_ctdf}
\title{Coerce an object to clusterTrack data format}
\usage{
as_ctdf(
  x,
  coords = c("longitude", "latitude"),
  time = "time",
  crs = NA,
  project_to,
  ...
)
}
\arguments{
\item{x}{A `data.table` object.}

\item{coords}{Character vector of length 2 specifying the coordinate column names.
Defaults to `c("longitude", "latitude")`.}

\item{time}{Name of the time column. Will be renamed to `"timestamp"` internally.}

\item{crs}{Coordinate reference system. Default is NA. Upstream methods will warn when unprojected
coordinates found (#TODO)}

\item{project_to}{passed to `st_transform()`.}

\item{...}{Currently unused}
}
\value{
An object of class `ctdf`, which inherits from `sf`.
}
\description{
Converts an object with spatial coordinates and a timestamp column
to a standardized `sf/data.table`-based format used internally by the clusterTrack package.
}
\note{
This is currently a thin wrapper around `st_as_sf()`, but standardizes timestamp naming, ordering,
and geometry column name (`"location"`). A colum `filter`, and with all values set to `FALSE` is added as well.
This column will be updated by upstream methods.
}
\examples{
data(zbird)
x = as_ctdf(zbird)

}
