% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdf.R
\name{as_ctdf}
\alias{as_ctdf}
\title{Coerce an object to clusterTrack data format}
\usage{
as_ctdf(x, ...)
}
\arguments{
\item{x}{An object to convert.}

\item{...}{Passed to methods.}
}
\value{
A \code{ctdf}.
}
\description{
S3 generic for converting objects into a \code{ctdf}.
}
