% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{as_tdbscan}
\alias{as_tdbscan}
\title{as_tdbscan}
\usage{
as_tdbscan(x, coords = c("longitude", "latitude"), time = "time", crs = 4326)
}
\arguments{
\item{x}{DT}

\item{coords}{default "longitude","latitude"}

\item{crs}{default 4326.}
}
\description{
as_tdbscan
}
\note{
For now just a thin wrapper on st_as_sf
}
\examples{
data(zbird)
x = as_tdbscan(zbird)
}
