% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_segments}
\alias{cluster_segments}
\title{Cluster movement tracks}
\usage{
cluster_segments(ctdf, threshold = 0.75, method = "quantile")
}
\arguments{
\item{ctdf}{A `ctdf`.}
}
\value{
NULL.  The `ctdf` object is updated in place.
}
\description{
Cluster movement tracks
}
\examples{
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2, crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf)
cluster_segments(ctdf )
map(ctdf)

data(lbdo66867)
ctdf = as_ctdf(lbdo66867, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf)
cluster_segments(ctdf)
map(ctdf)

data(pesa56511)
ctdf  = as_ctdf(pesa56511, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf) 
cluster_segments(ctdf) 
map(ctdf)

}
