% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_track.R
\name{print.clusterTrack}
\alias{print.clusterTrack}
\alias{cluster_segments}
\title{Cluster movement tracks}
\usage{
\method{print}{clusterTrack}(x, ...)

cluster_segments(ctdf)
}
\arguments{
\item{ctdf}{A `ctdf` object (from [as_ctdf()]), representing sequential movement points with timestamps and locations.}
}
\value{
A `ctdf` object containing only points assigned to clusters, with an additional integer column `cluster` indicating the cluster ID for each point (consecutive values starting from 1). Cluster membership reflects spatial-temporal groupings of points.
}
\description{
Perform clustering on a `ctdf` object to identify use-sites. Uses a hybrid approach combining Dirichlet-based polygon pruning and DBSCAN clustering to detect spatially and temporally coherent movement clusters ("use-sites").
}
\examples{
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2, crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf)
clust = cluster_segments(ctdf )
map(clust)

data(lbdo66867)
ctdf = as_ctdf(lbdo66867, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf)
clust = cluster_segments(ctdf)
map(clust)

data(pesa56511)
ctdf  = as_ctdf(pesa56511, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf) 
clust = cluster_segments(ctdf) 
map(clust)

}
