% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_track.R
\name{cluster_track}
\alias{cluster_track}
\title{Cluster movement tracks}
\usage{
cluster_track(
  ctdf,
  deltaT = 1,
  threshold = 0.75,
  method = "quantile",
  overlap_threshold = 0
)
}
\arguments{
\item{ctdf}{A `ctdf` data frame (see [as_ctdf()]) representing a single movement track, modified by reference.}

\item{deltaT}{Numeric. Passed to [slice_cdf()] The maximum temporal gap (in days) allowed between intersecting segments. Default is 1 day.}

\item{thereshold}{Numeric. ...}
}
\value{
NULL. The function modifies `ctdf` by reference, adding or updating the column \code{cluster}, which assigns a cluster ID to each row (point).
}
\description{
Performs spatial-temporal clustering on a `ctdf` object by identifying movement segments, then isolating their stop portions of a track and then and grouping them into clusters.
}
\details{
This is a high-level wrapper function that applies a pipeline of segmentation, clustering, and stitching steps on a movement track stored in a `ctdf` object.
}
\examples{
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2, s_srs = 4326, t_srs = "+proj=eqearth")
cluster_track(ctdf)

data(pesa56511)
ctdf  = as_ctdf(pesa56511, time = "locationDate", s_srs = 4326, t_srs = "+proj=eqearth")
cluster_track(ctdf)

data(lbdo66867)
ctdf = as_ctdf(lbdo66867, time = "locationDate", s_srs = 4326, t_srs = "+proj=eqearth")
cluster_track(ctdf)

data(lbdo66862)
ctdf = as_ctdf(lbdo66862, time = "locationDate", s_srs = 4326, t_srs = "+proj=eqearth")
cluster_track(ctdf)


}
