% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_track.R
\name{print.clusterTrack}
\alias{print.clusterTrack}
\alias{cluster_track}
\title{Cluster movement tracks}
\usage{
\method{print}{clusterTrack}(x, ...)

cluster_track(ctdf, sd = 1, transform = log, minPts = 5, borderPoints = TRUE)
}
\arguments{
\item{ctdf}{A `ctdf` object (from [as_ctdf()]), representing sequential movement points with timestamps and locations.}

\item{sd}{Numeric; threshold on the scaled log-area of Dirichlet polygons. Polygons with scaled log-area below this threshold are retained for neighbor calculation; expressed in standard deviations from the mean of the log-area.
Default = 1.}

\item{minPts}{Integer; minimum number of neighbors (including self) to form a core point in DBSCAN. Default = 5.}

\item{borderPoints}{Logical; if `FALSE`, treat all border points as noise (DBSCAN*). Default = TRUE.}
}
\value{
A `ctdf` object containing only points assigned to clusters, with an additional integer column `cluster` indicating the cluster ID for each point (consecutive values starting from 1). Cluster membership reflects spatial-temporal groupings of points.
}
\description{
Perform clustering on a `ctdf` object to identify use-sites. Uses a hybrid approach combining Dirichlet-based polygon pruning and DBSCAN clustering to detect spatially and temporally coherent movement clusters ("use-sites").
}
\examples{
library(clusterTrack)
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2, crs = 4326, project_to='+proj=eqearth')
slice_ctdf(ctdf)
o = cluster_track(ctdf)
plot(o)
plot(st_as_sf(ctdf)|>st_geometry(), add = TRUE)
s = st_as_sf(o)
mapview(s, zcol = "cluster")

}
