% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_track.R
\name{cluster_track}
\alias{cluster_track}
\title{Cluster movement tracks}
\usage{
cluster_track(
  ctdf,
  deltaT = 1,
  nmin = 3,
  threshold = 1,
  method = "sd",
  time_contiguity = FALSE,
  overlap_threshold = 0.1
)
}
\arguments{
\item{ctdf}{A `ctdf` data frame (see [as_ctdf()]) representing a single movement track .}

\item{deltaT}{Numeric; maximum allowable gap (in days) between segment endpoints to consider them continuous.
Default to 1 day. Passed to [slice()]) .}

\item{nmin}{Integer. Segments or tessellations with fewer than nmin points yield no clusters.
Default to 3. Passed to [cluster_segments()].}

\item{threshold}{Numeric. If `method = "sd"`, interpreted as the probability for
the area quantile used in pruning. Passed to [cluster_segments()].
If `method = "quantile"`, interpreted as the multiplier of the standard deviation
on log‐areas used in pruning. Passed to [cluster_segments()].}

\item{method}{Character, one of `"sd"` or `"quantile"`. Determines the pruning
strategy in the tessellation step:
- `"quantile"`: prune cells with area < quantile(polygon_area, probs = threshold)
- `"sd"`: prune cells with log(area) ≤ mean(log(area)) + threshold * sd(log(area)).
Passed to [cluster_segments()].}

\item{time_contiguity}{Logical; if `TRUE`, missing cluster IDs are forward‐filled.
and backward‐filled within each segment to enforce temporal continuity.
Default to `FALSE`.
Passed to [cluster_segments()].}

\item{overlap_threshold}{Numeric between 0 and 1; minimum area‐overlap ratio
required to merge adjacent clusters. Default to 0.1.
Clusters with overlap > threshold are combined.
Passed to [stitch_cluster()]}
}
\value{
NULL. The function modifies `ctdf` by reference, adding or updating the column \code{cluster}, which assigns a cluster ID to each row (point).
}
\description{
Performs spatiotemporal clustering on a ctdf by segmenting movement, identifying stops, and applying DBSCAN-like clustering.
}
\details{
This is a high-level wrapper function that applies a pipeline of segmentation, clustering, and stitching steps on a movement track stored in a `ctdf` object.
}
\examples{
data(toy_ctdf_k3)
ctdf = as_ctdf(toy_ctdf_k3)
cluster_track(ctdf)
\dontrun{
data(pesa56511)
ctdf = as_ctdf(pesa56511, time = "locationDate")
cluster_track(ctdf)

data(lbdo66862)
ctdf = as_ctdf(lbdo66862, time = "locationDate")
cluster_track(ctdf)


data(ruff143730)
ctdf = as_ctdf(ruff143730, time = "locationDate")
cluster_track(ctdf )


}
}
