% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{cluster_track}
\alias{cluster_track}
\title{Cluster movement tracks using DBSCAN}
\usage{
cluster_track(ctdf, minPts = 5, borderPoints = TRUE, ...)
}
\arguments{
\item{ctdf}{A `ctdf` object (from [as_ctdf()]).}

\item{minPts}{Integer; minimum number of neighbors (including self) to form a core point. Default = 5.}

\item{borderPoints}{Logical; if `FALSE`, treat all border points as noise (DBSCAN*). Default = TRUE.}

\item{...}{Passed to **dbscan::dbscan()**, currently unused.}
}
\value{
The input `ctdf`, invisibly, with a new integer column `cluster_id` of cluster labels (0 = noise).
}
\description{
Perform clustering on a `ctdf` object to identify use-sites.
}
\examples{
library(clusterTrack)
data(zbird)
ctdf = as_ctdf(zbird, time= 'locationDate')

filter_intersection(ctdf, overwrite = TRUE )
prune_delaunay_edges(ctdf, 0.5, 0.5)
cluster_track(ctdf)
plot(ctdf, by = 'cluster')

}
