% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_track.R
\name{cluster_track}
\alias{cluster_track}
\title{Cluster movement tracks}
\usage{
cluster_track(
  ctdf,
  deltaT = 30,
  threshold = 0.75,
  method = "quantile",
  overlap_threshold = 0
)
}
\arguments{
\item{ctdf}{A `ctdf` object (from [as_ctdf()]), representing sequential movement points with timestamps and locations.}
}
\value{
NULL.  The `ctdf` object is updated in place.
}
\description{
Perform clustering on a `ctdf` object to identify use-sites.
}
\examples{
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2, crs = 4326, project_to = "+proj=eqearth") |>cluster_track()
map(ctdf)
data(pesa56511)
ctdf  = as_ctdf(pesa56511, time = "locationDate", crs = 4326, project_to = "+proj=eqearth") |>cluster_track()
map(ctdf)

data(lbdo66867)
ctdf = as_ctdf(lbdo66867, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")|>cluster_track()
map(ctdf)

data(lbdo66862)
ctdf = as_ctdf(lbdo66862, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")|>cluster_track()
map(ctdf)


}
