% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTrack-package.R
\docType{package}
\name{clusterTrack-package}
\alias{clusterTrack}
\alias{clusterTrack-package}
\title{clusterTrack: Spatiotemporal clustering of animal telemetry tracks}
\description{
\code{clusterTrack} identifies spatiotemporally distinct use sites from animal
telemetry tracks.
}
\section{Package options}{

\code{clusterTrack} uses the following option:

\code{clusterTrack.verbose}: Logical. Default \code{TRUE}. Set to \code{FALSE} to
suppress \code{cli} alerts emitted during package workflows.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(clusterTrack.verbose = FALSE)
}\if{html}{\out{</div>}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://ornitho-logics.github.io/clusterTrack/}
  \item Report bugs at \url{https://github.com/ornitho-logics/clusterTrack/issues}
}

}
\author{
\strong{Maintainer}: Mihai Valcu \email{mvalcu@gwdg.de} (\href{https://orcid.org/0000-0002-6907-7802}{ORCID})

Authors:
\itemize{
  \item Mihai Valcu \email{mvalcu@gwdg.de} (\href{https://orcid.org/0000-0002-6907-7802}{ORCID})
}

}
\keyword{internal}
