% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessellate.R
\name{.tesselate}
\alias{.tesselate}
\title{ctdf = as_ctdf(toy_ctdf_k3,s_srs = 4326, t_srs = "+proj=eqearth") |> slice_ctdf()
x = ctdf[.segment == 1]
 plot(tess); plot(env, add = TRUE)}
\usage{
.tesselate(x)
}
\description{
ctdf = as_ctdf(toy_ctdf_k3,s_srs = 4326, t_srs = "+proj=eqearth") |> slice_ctdf()
x = ctdf[.segment == 1]
 plot(tess); plot(env, add = TRUE)
}
