% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_ruff}
\alias{mini_ruff}
\title{Simulated ARGOS tracking data with 3 spatial clusters}
\format{
A data.table with 44 rows and 3 columns:
\describe{
\item{longitude}{Numeric. Longitude in decimal degrees (WGS84).}
\item{latitude}{Numeric. Latitude in decimal degrees (WGS84).}
\item{time}{POSIXct. Timestamp of location fix (UTC).}
}
}
\usage{
mini_ruff
}
\description{
A toy dataset simulating ARGOS satellite tracking data for one individual.
The dataset contains timestamped locations arranged in three distinct spatial clusters.
}
\details{
This dataset is fully synthetic and was created to represent idealized movement
within and between three clusters. There is no associated individual or species.
}
\examples{
data(mini_ruff)
plot(mini_ruff$longitude, mini_ruff$latitude, type = "l")
}
\keyword{datasets}
