% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_delaunay_edges}
\alias{prune_delaunay_edges}
\title{Prune Delaunay Triangulation and Extract Pruned Edges}
\usage{
prune_delaunay_edges(ctdf, global_sd = 1, local_sd = 1)
}
\arguments{
\item{ctdf}{A `ctdf` data frame.}

\item{global_sd}{Numeric multiplier for the global threshold (default = 1).}

\item{local_sd}{Numeric multiplier for the local threshold (default = 1).}
}
\value{
Invisibly returns the input \code{ctdf} object which got a new \code{neighbors} column.
}
\description{
Given a `ctdf`  this function performs:
1. Delaunay triangulation via `deldir`.
2. Global-effect removal: discards triangles whose area or longest side 
   exceeds (mean + global_sd * sd).
3. Local-effect removal: among survivors, discards triangles whose longest 
   side exceeds (mean + local_sd * sd) of remaining edges.
4. Builds the adjacency edge list of remaining triangles.
}
\note{
TODO: add reference
}
\examples{
library(clusterTrack)
data(zbird)
ctdf = as_ctdf(zbird)
filter_intersection(ctdf)
prune_delaunay_edges(ctdf)
plot(ctdf, by = 'filter')

}
