% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{putative_cluster_trace}
\alias{putative_cluster_trace}
\title{Extract putative-cluster trace}
\usage{
putative_cluster_trace(x)
}
\arguments{
\item{x}{A \code{ctdf} object returned by \code{cluster_track(trace = TRUE)}.}
}
\value{
A \code{data.table}, or \code{NULL}.
}
\description{
Extract putative-cluster trace
}
\examples{
data(mini_ruff)

x = as_ctdf(mini_ruff)
cluster_track(x, trace = TRUE)

tr = putative_cluster_trace(x)
tr
}
