% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ruff143789}
\alias{ruff143789}
\title{ARGOS satellite tracking data for an individual Ruff}
\format{
A data.table with 2834 rows and 4 columns:
\describe{
  \item{latitude}{Numeric. Latitude in decimal degrees (WGS84).}
  \item{longitude}{Numeric. Longitude in decimal degrees (WGS84).}
  \item{locationDate}{POSIXct. Timestamp of location fix (UTC).}
  \item{locationClass}{Character. ARGOS location quality class (e.g., B, 0, A).}
}
}
\source{
Kempenaers, Bart, et al. "Large-scale sampling of potential breeding sites in male ruffs."
         Proceedings B 292.2038 (2025): 20242225.
         https://royalsocietypublishing.org/doi/full/10.1098/rspb.2024.2225.
}
\usage{
ruff143789
}
\description{
A dataset containing ARGOS locations for an individual male Ruff.
}
\examples{
data(ruff143789)
head(ruff143789)
}
\keyword{datasets}
