% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_ctdf}
\alias{slice_ctdf}
\title{segment and filter a ctdf
slice_ctdf}
\usage{
slice_ctdf(ctdf, deltaT = 30)
}
\description{
segment and filter a ctdf
slice_ctdf
}
\examples{
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2, s_srs = 4326, t_srs = "+proj=eqearth")
ctdf = slice_ctdf(ctdf)  

data(pesa56511)
ctdf = as_ctdf(pesa56511, time = "locationDate", s_srs = 4326, t_srs = "+proj=eqearth")
slice_ctdf(ctdf )   
}
