% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_ctdf}
\alias{slice_ctdf}
\title{segment and filter a ctdf
slice_ctdf}
\usage{
slice_ctdf(
  ctdf,
  deltaT = 12,
  slice_min_n_segments = 5,
  slice_min_bout_length,
  filter_min_n_segments = 5
)
}
\description{
segment and filter a ctdf
slice_ctdf
}
\examples{
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2, crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf )
se = st_as_sf(ctdf)
mapviewOptions(fgb = FALSE)
mapview(se, zcol = ".segment")
mapview(se, zcol = ".filter")

}
