% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{slice_ctdf}
\alias{slice_ctdf}
\title{Segment and filter a CTDF by temporal continuity and spatial clustering}
\usage{
slice_ctdf(ctdf, deltaT = 1)
}
\arguments{
\item{ctdf}{A CTDF object.}

\item{deltaT}{Numeric; maximum allowable gap (in days) between segment
endpoints to consider them continuous.}

\item{progress_bar}{Logical; whether to display a progress bar during execution. Defaults to `TRUE`.}
}
\value{
The input CTDF, updated (in-place) with an integer
  \code{.putative_cluster} column indicating bout membership.
}
\description{
Recursively splits a CTDF into continuous bouts. The split stops when any bout  has one
cluster (via HDBSCAN).
}
\examples{
data(toy_ctdf_k3)
ctdf = as_ctdf(toy_ctdf_k3, s_srs = 4326, t_srs = "+proj=eqearth")
ctdf = slice_ctdf(ctdf)

data(pesa56511)
ctdf = as_ctdf(pesa56511, time = "locationDate", s_srs = 4326, t_srs = "+proj=eqearth")
slice_ctdf(ctdf )
}
