% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_slice.R
\name{slice_ctdf}
\alias{slice_ctdf}
\title{Segment and filter a CTDF by temporal continuity and spatial clustering}
\usage{
slice_ctdf(ctdf, deltaT = 30, nmin = 5)
}
\arguments{
\item{ctdf}{A CTDF object.}

\item{deltaT}{Numeric; maximum allowable gap (in days) between segment
endpoints to consider them continuous.}
}
\value{
The input CTDF, updated (in-place) with an integer
  \code{.putative_cluster} column indicating bout membership.
}
\description{
Recursively splits a CTDF into putative cluster regions. The split stops when a region is homogenous
 (established via HDBSCAN).
}
\examples{
data(mini_ruff)
ctdf = as_ctdf(mini_ruff, s_srs = 4326, t_srs = "+proj=eqearth")
ctdf = slice_ctdf(ctdf)

data(pesa56511)
ctdf = as_ctdf(pesa56511, time = "locationDate", s_srs = 4326, t_srs = "+proj=eqearth")
slice_ctdf(ctdf )
}
