% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch_cluster}
\alias{stitch_cluster}
\title{stitch_cluster
stitch a cluster}
\usage{
stitch_cluster(clust, overlap_threshold = 0)
}
\description{
stitch_cluster
stitch a cluster
}
\examples{
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2, crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf)
clust = cluster_segments(ctdf )
map(clust)
data(pesa56511)
ctdf  = as_ctdf(pesa56511, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf) 
clust = cluster_segments(ctdf)
clust = stitch_cluster(clust)
map(clust)

data(lbdo66867)
ctdf = as_ctdf(lbdo66867, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf)
clust = cluster_segments(ctdf)
clust =  stitch_cluster(clust)
map(clust)
}
