% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch_cluster}
\alias{stitch_cluster}
\title{stitch_cluster
stitch a cluster}
\usage{
stitch_cluster(ctdf, overlap_threshold = 0)
}
\description{
stitch_cluster
stitch a cluster
}
\examples{
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2, crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf)
cluster_segments(ctdf)
stitch_cluster(ctdf)
map(ctdf)
data(pesa56511)
ctdf  = as_ctdf(pesa56511, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf) 
cluster_segments(ctdf)
stitch_cluster(ctdf)
map(ctdf)

data(lbdo66867)
ctdf = as_ctdf(lbdo66867, time = "locationDate", crs = 4326, project_to = "+proj=eqearth")
slice_ctdf(ctdf)
cluster_segments(ctdf)
stitch_cluster(ctdf)
map(ctdf)
}
