% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessellate.R
\name{tessellate_ctdf}
\alias{tessellate_ctdf}
\title{Tesselate a ctdf}
\usage{
tessellate_ctdf(ctdf)
}
\arguments{
\item{ctdf}{A `ctdf` data frame.}
}
\value{
un updated ctdf object.
}
\description{
This function computes Dirichlet (Voronoi) polygons on each putative_cluster
of a `ctdf` object
}
\examples{
library(clusterTrack)
data(mini_ruff)
ctdf = as_ctdf(mini_ruff,s_srs = 4326, t_srs = "+proj=eqearth") |> slice_ctdf()
tessellate_ctdf(ctdf)

}
