% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessellate.R
\name{tessellate_ctdf}
\alias{tessellate_ctdf}
\title{Tesselate a ctdf}
\usage{
tessellate_ctdf(ctdf)
}
\arguments{
\item{ctdf}{A `ctdf` data frame (typically created with [as_ctdf()]).}
}
\value{
A sf object.
}
\description{
This function computes Dirichlet (Voronoi) polygons from a `ctdf` object
}
\note{
Dirichlet polygons are computed via the `deldir` package.
}
\examples{
library(clusterTrack)
data(toy_ctdf_k2)
ctdf = as_ctdf(toy_ctdf_k2,crs = 4326, project_to = "+proj=eqearth")
x = tessellate_ctdf(ctdf)
s = prune_tesselation(x, q = 0.90)
poly2nb(s, queen = TRUE)

}
\seealso{
[as_ctdf()]
}
