% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{track_segments}
\alias{track_segments}
\title{Convert a `ctdf` track to movement step segments as LINESTRINGs}
\usage{
track_segments(ctdf)
}
\arguments{
\item{ctdf}{A `ctdf` object (with ordered rows and a `"location"` geometry column).}
}
\value{
An `sf` object with LINESTRING geometry for each step.
}
\description{
Takes a `ctdf` object and returns an `sf` object with LINESTRING geometries representing
the movement steps between consecutive locations. Each segment connects two points,
starting at the previous location and ending at the current one - i.e., each segment
ends at the position of the current row.
}
\details{
The output includes additional columns for step duration (`step_duration`), 
step distance (`step_distance` in meters), and speed (`speed` in km/h). 
The number of rows is nrow(ctdf) - i, where i = 1 and corresponds to the starting index in ctdf.
}
\examples{
data(pesa56511)
ctdf = as_ctdf(pesa56511, time = 'locationDate')
s = track_segments(ctdf)
plot(s[c('timestamp', 'speed')])

}
