% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{track_segments}
\alias{track_segments}
\title{Convert `ctdf` point track to step segments as LINESTRINGs}
\usage{
track_segments(x)
}
\arguments{
\item{ctdf}{A `ctdf` object (or any `sf` POINT object with time-ordered rows).}
}
\value{
An `sf` object with LINESTRING geometry for each step.
}
\description{
Takes a `ctdf` object and returns
an `sf` object of LINESTRINGs representing the movement steps between
consecutive locations.
}
\examples{
data(pesa56511)
x = as_ctdf(pesa56511, time = 'locationDate')
segs = ctdf_to_segments(x)
plot(segs)

}
